requirements.txt
asyncpg==0.29.0
ib_insync==0.9.86
ibapi==9.81.1.post1
pandas==2.2.2
utils==1.0.2
numpy==1.26.4
psycopg2==2.9.9
pg8000==1.29.8
SQLAlchemy==2.0.17
cloud-sql-python-connector[pg8000]
google-api-core==2.17.1
google-auth==2.26.2
google-cloud==0.34.0
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0